requirements.txt
django==3.2.7
djangorestframework==3.11.2
Markdown==3.2.1
Pillow==8.3.2
django-imagekit
sorl-thumbnail==12.7.0
coreapi==2.3.3
psycopg2>=2.7,<3.0
